requirements.txt
phidata 
python-dotenv
yfinance
packaging
duckduckgo-search
spider-client
fastapi
uvicorn
groq
openai
streamlit
